requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==24.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
exponent_server_sdk==2.1.0
fastapi==0.110.3
firebase-admin==6.6.0
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
lxml==5.3.0
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
passlib==1.7.4
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
rollbar==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-cockroachdb==2.0.2
sqlalchemy-data-model-visualizer==0.1.3
starlette==0.37.2
tenacity==9.0.0
twilio==9.3.7
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.27.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
yarl==1.17.1
